# Cache projections in trans