… src/librustc_trans/debuginfo/mod.rs
… src/librustc_trans/glue.rs
… src/librustc_trans/meth.rs
… src/librustc_trans/mir/constant.rs
… src/librustc_trans/monomorphize.rs
… src/librustc_trans/partitioning.rs
… src/librustc_trans/trans_item.rs