@nikomatsakis 2016-08-31
(I'll rebase shortly.)