@eddyb 2016-08-31
r=me modulo `closure_type`.